Give an Operation's state one number, not two
A state travelled twice and by two different numberings. The state itself had one, given by its place in
the list; the log a world saves and the packets a terminal reads had another, written out separately and
in a different order. Nothing joined them and nothing would have said so: add a state to one, or move one,
and every log already saved reads as something else.

There is one now, carried by the state, starting at one so that zero is free to mean that nothing is known
about an Operation rather than meaning the first of the eight. A switch needs a label the compiler can
read, so the log still writes the numbers out; a test fails the moment one of them and the state it names
stop agreeing, which is the part the compiler cannot hold.

Priorities are numbered in tens instead of one after another, so a level can be put between two of them
later without any number already written into a world having to move.

How long an Operation waited and how long it ran are counted without a ceiling.

What an Operation requires of a network is copied on the way in and on the way out, so that reading what
it needs can no longer be a way of changing it.

+++ computers/src/main/java/dev/jstech/computers/blockentity/MainframeBlockEntity.java
@@ -1737,7 +1737,7 @@ public class MainframeBlockEntity extends AbstractComputerBlockEntity
                              final OperationRecord record) {
         if (record.completed()) {
             post(net -> new IOperationLifecycleEvent.Completed(net,
-                    operation.operationId(), operation.typeId(), (long) record.waitedTicks() + record.ranTicks()));
+                    operation.operationId(), operation.typeId(), record.waitedTicks() + record.ranTicks()));
         } else if (record.status() == OperationRecord.STATUS_DISCARDED) {
             post(net -> new IOperationLifecycleEvent.Discarded(net,
                     operation.operationId(), operation.typeId()));
